+++ addons/mass_mailing/security/mass_mailing_security.xml
@@ -6,12 +6,6 @@
         <field name="users" eval="[(4, ref('base.user_root'))]"/>
         <field name="implied_ids" eval="[(4, ref('base.group_user'))]"/>
     </record>
-    
-
-    <record id="group_website_popup_on_exit" model="res.groups">
-        <field name="name">Use subscription pop up on the website</field>
-        <field name="category_id" ref="base.module_category_hidden"/>
-    </record>
 
     <!-- Group to manage campaigns -->
     <record id="group_mass_mailing_campaign" model="res.groups">
