+++ addons/mass_mailing/views/mass_mailing_views.xml
@@ -154,15 +154,7 @@
                             <h1>
                                 <field name="name" placeholder="e.g. Consumer Newsletter"/>
                             </h1>
-                            <group groups="mass_mailing.group_website_popup_on_exit">
-                                <field name="popup_redirect_url"/>
-                            </group>
                         </div>
-                        <notebook groups="mass_mailing.group_website_popup_on_exit">
-                            <page string="Popup Content">
-                                <field name="popup_content" widget="html_frame" options="{'editor_url': '/website_mass_mailing/field/popup_content'}"/>
-                            </page>
-                        </notebook>
                     </sheet>
                 </form>
             </field>
