+++ addons/website_mass_mailing/__openerp__.py
@@ -10,9 +10,12 @@ Add a snippet in the website builder to subscribe a mass_mailing list
     'category': 'Marketing',
     'depends': ['website', 'mass_mailing'],
     'data': [
+        'security/mass_mailing_security.xml',
         'views/website_mass_mailing_templates.xml',
         'views/unsubscribe_templates.xml',
         'views/snippets_templates.xml',
+        'views/mass_mailing_view.xml',
+        'views/res_config_view.xml',
     ],
     'auto_install': True,
 }
