+++ addons/website_mass_mailing/models/__init__.py
@@ -1,5 +1,5 @@
 # -*- coding: utf-8 -*-
 # Part of Odoo. See LICENSE file for full copyright and licensing details.
 
-from . import controllers
-from . import models
+from . import mass_mailing_list
+from . import res_config
