+++ addons/website_mass_mailing/security/mass_mailing_security.xml
@@ -1,0 +1,7 @@
+<?xml version="1.0" encoding="utf-8"?>
+<odoo>
+    <record id="group_website_popup_on_exit" model="res.groups">
+        <field name="name">Use subscription pop up on the website</field>
+        <field name="category_id" ref="base.module_category_hidden"/>
+    </record>
+</odoo>
