+++ addons/website_mass_mailing/views/res_config_view.xml
@@ -1,0 +1,13 @@
+<?xml version="1.0" encoding="utf-8"?>
+<odoo>
+    <record id="inherited_view_mass_mailing_configuration" model="ir.ui.view">
+        <field name="inherit_id" ref="mass_mailing.view_mass_mailing_configuration"/>
+        <field name="name">mass.mailing.config.settings.inherit</field>
+        <field name="model">mass.mailing.config.settings</field>
+        <field name="arch" type="xml">
+            <xpath expr="//group[@name='mass_mailing_group']" position="inside">
+                <field name="group_website_popup_on_exit" widget="radio"/>
+            </xpath>
+        </field>
+    </record>
+</odoo>
