+++ addons/website_mass_mailing/views/snippets_templates.xml
@@ -38,7 +38,7 @@
 
 <template id="website_popup_snippet" name="popup snippet" inherit_id="website.snippets">
     <xpath expr="//div[@id='snippet_feature']/div[@class='o_panel_body']" position="inside">
-        <div groups="mass_mailing.group_website_popup_on_exit">
+        <div groups="website_mass_mailing.group_website_popup_on_exit">
             <div class="oe_snippet_thumbnail">
                 <img class="oe_snippet_thumbnail_img" src="/mass_mailing/static/src/img/blocks/bounce.png" />
                 <span class="oe_snippet_thumbnail_title">Newsletter Popup</span>
